+++ src/router/GuestRouter.tsx
@@ -1,6 +1,7 @@
 import { Routes, Route } from 'react-router-dom';
 
-import GuestLogin from 'pages/Guest/GuestLogin';
+import GuestLogin from 'pages/Guest/GuestOrder';
+import GuestOrder from 'pages/Guest/GuestOrder';
 import GuestOrderDetail from 'pages/Guest/GuestOrderDetail';
 import NotFound from 'pages/NotFound';
 
@@ -8,6 +9,7 @@ const GuestRouter = () => {
     return (
         <Routes>
             <Route path='login' element={<GuestLogin />} />
+            <Route path='order' element={<GuestOrder />} />
             <Route
                 path='order/detail/:guestOrderNo'
                 element={<GuestOrderDetail />}
